Corrección estructura
Se eliminaron archivos duplicados y se generó la carpeta components en donde contiene los .js de las vistas creadas.
