+++ src/main.js
@@ -1,7 +1,7 @@
 // Este es el punto de entrada de tu aplicacion
-import { Home } from './Home.js';
-import { Register } from './Register.js';
-import { Login } from './Login.js';
+import { Home } from './components/home.js';
+import { Register } from './components/register.js';
+import { Login } from './components/login.js';
 
 const rootDiv = document.getElementById('root');
